Add site-branding style for sitename & digest (& avatar, etc)
Display nav menu items in a row below site-branding. They are not visible in portrait orientation due to removal of the hamburger button that toggles the show state.

Next revisions will be to select between nav menu (when window is wide enough) and burger menu (when window is too narrow).